# socks-client/swagger_client/models/sock_request.py
# coding: utf-8

"""
    OpenAPI definition

    No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)  # noqa: E501

    OpenAPI spec version: v0
    
    Generated by: https://github.com/swagger-api/swagger-codegen.git
"""

import pprint
import re  # noqa: F401

import six

class SockRequest(object):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    """
    Attributes:
      swagger_types (dict): The key is attribute name
                            and the value is attribute type.
      attribute_map (dict): The key is attribute name
                            and the value is json key in definition.
    """
    swagger_types = {
        'color': 'str',
        'cotton_part': 'int',
        'quantity': 'int'
    }

    attribute_map = {
        'color': 'color',
        'cotton_part': 'cottonPart',
        'quantity': 'quantity'
    }

    def __init__(self, color=None, cotton_part=None, quantity=None):  # noqa: E501
        """SockRequest - a model defined in Swagger"""  # noqa: E501
        self._color = None
        self._cotton_part = None
        self._quantity = None
        self.discriminator = None
        self.color = color
        if cotton_part is not None:
            self.cotton_part = cotton_part
        if quantity is not None:
            self.quantity = quantity

    @property
    def color(self):
        """Gets the color of this SockRequest.  # noqa: E501


        :return: The color of this SockRequest.  # noqa: E501
        :rtype: str
        """
        return self._color

    @color.setter
    def color(self, color):
        """Sets the color of this SockRequest.


        :param color: The color of this SockRequest.  # noqa: E501
        :type: str
        """
        if color is None:
            raise ValueError("Invalid value for `color`, must not be `None`")  # noqa: E501

        self._color = color

    @property
    def cotton_part(self):
        """Gets the cotton_part of this SockRequest.  # noqa: E501


        :return: The cotton_part of this SockRequest.  # noqa: E501
        :rtype: int
        """
        return self._cotton_part

    @cotton_part.setter
    def cotton_part(self, cotton_part):
        """Sets the cotton_part of this SockRequest.


        :param cotton_part: The cotton_part of this SockRequest.  # noqa: E501
        :type: int
        """

        self._cotton_part = cotton_part

    @property
    def quantity(self):
        """Gets the quantity of this SockRequest.  # noqa: E501


        :return: The quantity of this SockRequest.  # noqa: E501
        :rtype: int
        """
        return self._quantity

    @quantity.setter
    def quantity(self, quantity):
        """Sets the quantity of this SockRequest.


        :param quantity: The quantity of this SockRequest.  # noqa: E501
        :type: int
        """

        self._quantity = quantity

    def to_dict(self):
        """Returns the model properties as a dict"""
        result = {}

        for attr, _ in six.iteritems(self.swagger_types):
            value = getattr(self, attr)
            if isinstance(value, list):
                result[attr] = list(map(
                    lambda x: x.to_dict() if hasattr(x, "to_dict") else x,
                    value
                ))
            elif hasattr(value, "to_dict"):
                result[attr] = value.to_dict()
            elif isinstance(value, dict):
                result[attr] = dict(map(
                    lambda item: (item[0], item[1].to_dict())
                    if hasattr(item[1], "to_dict") else item,
                    value.items()
                ))
            else:
                result[attr] = value
        if issubclass(SockRequest, dict):
            for key, value in self.items():
                result[key] = value

        return result

    def to_str(self):
        """Returns the string representation of the model"""
        return pprint.pformat(self.to_dict())

    def __repr__(self):
        """For `print` and `pprint`"""
        return self.to_str()

    def __eq__(self, other):
        """Returns true if both objects are equal"""
        if not isinstance(other, SockRequest):
            return False

        return self.__dict__ == other.__dict__

    def __ne__(self, other):
        """Returns true if both objects are not equal"""
        return not self == other